High: pgsql: add a new attribute as $PGSQL_CHILDREN and attribute value as HS:cascade.
On the existing pgsql RA, all slave nodes are master node's children.
But in cascade replication, it is possible that a slave node have it's children.
Therefor, it is necessary to manage that a node is whose child, and it is specified by $PGSQL_CHILDREN.
In addition, when the 3 or more nodes are on cascade, HS:cascade is shown instead of HS:async.