High: pgsql: follow the specification change of "synchronous_standby_names" on PostgreSQL 9.6
This patch fix the following issue on PostgreSQL 9.6
On PostgreSQL 9.6 the 'synchronous_standby_names' changed to SQL identifier from plain text.
Now 'synchronous_standby_names' and 'application_name' is made from node name by pgsql RA.
There are some different rules between the node name and the SQL idenfier.
As a result, if the node name include "-", "." and the first character is number, the node
can not be allowed synchronous replication.